requirements.txt
#enconding: utf-8
Django==2.0.2
django-constance==2.1.0
django-extensions==1.9.9
django-filter==1.1.0
django-picklefield==1.0.0
django-reversion==2.0.12
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.1
djangorestframework==3.9.0
easy-thumbnails==2.5
hashids==1.2.0
pytz==2017.3